moved corrected increment math into digit index object

+++ FlipCounter/FlipCounterView.h
@@ -25,10 +25,9 @@
     BOOL isAnimating;
 }
 
-- (void) add:(NSUInteger)incr;
+- (void) add:(float)incr;
 
 @end
-
 
 
 @interface FlipCounterViewDigitIndex : NSObject
